requirements.txt
numpy>=1.20
rasterio>=1.2
gdown>=5.0
torch>=1.9
google-cloud-storage>=2.0
google-cloud-auth>=2.0
